requirements.txt
watchdog -> api used to talk with my pc
pillow -> image processing library
krita-python -> krita python api to convert kra to jpg